move cart modal from header to shopping page, fix transitions for CategoryItemsGrid
